requirements.txt
colormath>=3.0.0
UnityPy>=1.10
leb128>=1.0.5
pillow>=10.0.0
xmltodict
